requirements.txt
twisted==22.10.0
mysql-connector-python==8.0.32